# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a database replay tool API server.",
        "title": "RUC DB Replay API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/replay/prepare": {
            "post": {
                "description": "上传采集日志文件和目标数据库连接信息，解析日志并存储",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Replay"
                ],
                "summary": "准备回放环境",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Database Host",
                        "name": "db_host",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Database Port",
                        "name": "db_port",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Database User",
                        "name": "db_user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Database Password",
                        "name": "db_pass",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Database Name",
                        "name": "db_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "PgAudit Log File",
                        "name": "log_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.PrepareResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/replay/progress": {
            "get": {
                "description": "查询回放任务的当前进度",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Replay"
                ],
                "summary": "获取回放进度",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/replay/report": {
            "get": {
                "description": "查询回放任务的执行报告",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Replay"
                ],
                "summary": "获取回放报告",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.ReportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/replay/run": {
            "post": {
                "description": "开始回放任务，返回任务状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Replay"
                ],
                "summary": "启动流量回放",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID returned from prepare",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.RunResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/replay/statements": {
            "get": {
                "description": "分页获取任务的SQL语句",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Replay"
                ],
                "summary": "获取SQL语句列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/replay/task": {
            "get": {
                "description": "查询任务的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Replay"
                ],
                "summary": "获取任务详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/replay/transactions": {
            "get": {
                "description": "获取任务的所有事务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Replay"
                ],
                "summary": "获取事务列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/replay/tx/statements": {
            "get": {
                "description": "获取指定事务的所有SQL语句",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Replay"
                ],
                "summary": "获取事务的SQL语句",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "tx_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.PrepareResponse": {
            "type": "object",
            "properties": {
                "statistics": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "total_statements": {
                    "type": "integer"
                },
                "total_transactions": {
                    "type": "integer"
                }
            }
        },
        "handler.ReportResponse": {
            "type": "object",
            "properties": {
                "avg_latency_ms": {
                    "type": "number"
                },
                "duration": {
                    "type": "string"
                },
                "executed_stmts": {
                    "type": "integer"
                },
                "failed_stmts": {
                    "type": "integer"
                },
                "success_rate": {
                    "type": "number"
                },
                "success_stmts": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "string"
                },
                "total_statements": {
                    "type": "integer"
                },
                "total_transactions": {
                    "type": "integer"
                }
            }
        },
        "handler.RunResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "running"
                },
                "task_id": {
                    "type": "string",
                    "example": "task-12345"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}